requirements.txt
discord.py==2.2.2
praw==7.7.0
PyNaCl==1.5.0
deep_translator==1.10.0
wikipedia==1.4.0
lxml==4.9.2
aiosqlite==0.18.0
beautifulsoup4==4.11.2
revChatGPT==4.1.3
EdgeGPT==0.1.22.1
openai==0.27.2
googlebard==0.0.6.1
replicate==0.6.1